Make -sil-irgen default everywhere (except libs).
Everything can compile and run with SILGen now, but there are a handful of crashing tests when linking against a stdlib built with SIL, so make SWIFT_SIL_IRGEN a CMake option and disable it by default for now. Also tag the IRGen tests with -no-sil-irgen so that I don't have to rewrite their CHECK: lines right now.

Swift SVN r4749