Remove @_transparent references to allow better backward compatability
Older versions of Xcode that do not allow the hot-swappability of Swift
toolchains do not have the definition of @_transparent. The geometry
APIs are not fully clear if they are reasonable to live in Foundation
in the first place; so in the spirit of compatability we can just
remove that attribute for now.