Make sure that we include the correct version of CF when building Foundation, TestFoundation, and plutil
On OS X, if we are not careful about using the correct include paths then the compiler will pick up the system (SDK) version of CF instead of the one we just built as part of the project. This can lead to mismatches in modules, which will either cause a compiler error or crash.