Repair NSNumber behaviour
During initalization store the number type in the bit-field in the
same way that CFNumberCreate does-so.

NSNumber/CFNumber stores its value in a UInt64 (_pad) with associated type
information as part of the CFInfo bitfield. Initialization is performed
through _CFNumberInit which copies the value bytes into _pad but does
not store the type information.