Remove objective-c implicit bridging so that both linux and mac targets have pairity for conversions
This makes any translation between NSObject types and their structural counterparts need to be explicitly converted instead of implicity converted like they are in the objective-c interop version of Swift accessing Foundation on Darwin. Linux does not currently have this feature and it should help maintain a proper level of parity between the two platforms.