CoreFoundation: use C99 macro expansion
Rather than defining the various flavours of CFAssert requiring the user to
count the number of substituted parameters, collapse them to CFAssert.  The C99
variadic macro expansion permits us to ensure that the arguments are all
passable.  At the same time, the compiler validation of the printf conversion
will still catch the missing arguments (which could still be possible to have in
the previous case by just using the wrong count).