Fix a macOS test failure in TestHTTPCookie. (#523)
The root cause of the problem is missing logic in the CFTimeZone.c file
to handle "GMT" (no offset) as a time zone, which Darwin Foundation
handles. Also, CFLocale.c was using an empty string for the current locale, due
to not checking the AppleLocale preference. For now, we hard code en_US
as the current locale, although obviously this should be fixed to use
CFPreferences if possible.