Updating PR 299 (NSURLSession) to work with the latest Foundation (#426)
Also includes support for downloadTask and uploadTask with Data and a completion handler.
The original PR 299 doesn't build on Ubuntu 14.04, due to an older level of libcurl. This
rework removes some of that code (it does not affect any functionality). Updated to use the
Dispatch 3.0 API. Includes basic tests with external URLs.