Add initialisers and test coverage for Decimal
The Decimal type on Darwin is a packed structure, and so the
implementation in Swift needs to follow the same packed layout
to ensure that the values are truncated in the same way.

Add test coverage for initialisers and verify the same test
code passes on Swift on Darwin with the same results.