Final implementation of Decimal (#687)
* Final implementation of Decimal

This extends the normalisation routines to be able to handle
mismatched size elements, and division of decimals that are larger
than short values correctly. It adds tests that are the same coverage
as for NSDecimal on Darwin.

* Implement NSDecimalNumber

Using the functionality in NSDecimal, implement the NSDecimalNumber
such that the additional tests in TestNSDecimal pass.