Rename function to clarify behaviour and fix bug
The decimalCompare function could be confused with the NSDecimalCompare
function, but it does something very different. Since this function
only compares the mantissa, rename it to compareMantissa instead.

Also fix a bug which slipped through whereby the mantissa would be
compared in the wrong order, leading to some incorrect results when
comparing Decimal values.