Implement the remainder of NSDecimalNumber
The scanning functionality in NSDecimalNumber fits better in the
Decimal file, since it can access internal implementation details
in a better way. Implement it with the multiplyBy10:andAdd function
and use a similar pattern to the existing parseDouble method, but
using NSDecimal multiplication instead.

Add tests to verify new behaviour.