Update for new rules on @inline(__always) function bodies
Referencing private or internal declarations from an @inline(__always)
body can lead to SIL verifier and linker failures, so now we diagnose
and prohibit it.

Mark the entity in question public, or if it is not part of user-visible
API, internal and @_versioned.

See <https://github.com/apple/swift/pull/6669>.