re-enable building without libblocksruntime-dev (Blocks.h)
Add -I./closure to foundation.CFLAGS so that Block.h
can be found from closure/Block.h instead of assuming
that it will be available in the system includes.
This simplifies building Foundation on platforms where
libblocksruntime-dev is not easily available as a
pre-built package.

Also include Block.h in CFError.c to fix compilation warning.