Disable tests that use the internet
Plist DTDs were recently moved from
http://www.apple.com/DTDs/PropertyList-1.0.dtd to
https://www.apple.com/DTDs/PropertyList-1.0.dtd (note the https).
libXML, however, is not setup to request DTDs from secure domains and
fails.  This causes a spurious failure in the validation test checking
for failure and a real failure in the validation test checking for
success.  In either case, these tests introduce a hard dependency on
the internet into the test suite.  I’ve disabled them for now and filed
rdar://31567922 about this.