Small cleanup: Use () -> ... instead of (Void) -> ... for no-argument closures
In Swift, (Void) -> Void is actually a function that takes
a single value of empty-tuple type, and not a function that
takes no arguments.

Swift 3 largely ignored the distinction because of the implicit
'tuple splat' behavior; in Swift 4, this behavior has been
eliminated as part of implementing SE-0110, so calling a function
of type (Void) -> () now requires passing in an empty tuple,
like foo(()).

I believe this is not intended behavior, so this patch changes
the functions in question to be written as () -> Void.