Add support for currentDirectoryPath in Process class.
This patch adds support for the `currentDirectoryPath` property on `Process`

Since `posix_spawn` doesn't make it easy to change the working directory, a
call to `chdir` is issued just before the process executes. It looks like
`lldb` follows a similar pattern in one of its tools
(`tools/darwin-debug/darwin-debug.cpp`).