Replace precondition if with guard in NSNumber.encode.
Previously NSNumber.encode would call NSUnimplemented when you try to
do a keyed coding. Whilst it is true that this wasn't implemented it was
due to not being possible rather than not yet implemented.

With this patch we call preconditionFailure which is more consistent
with the rest of the project. We also use a guard so can remove a level of
indentation.