Provide an NSUnsupported macro for not supported implementations
When developing cross-platform applications, an `NSUnimplemented()`
is currently used to identify when a method is not and will not be
implemented.

Since `NSUnimplemented()` is used to also mean 'not yet implemented',
using a distinct name for the same functionality will allow the
two states to be distinguished, in the same way that
`NSRequiresConcreteImplementation()` is used to indicate that a
subtype must provide the behaviour.