Create StringMap.swift
This is a fast way to search and match strings from a String Array

This implementation is mainly improving the time taken to match a string from a string array. This isn't a complex or long implementation like other implementations, yet drastically reduces the time taken to do such searching and matching tasks. 

This is a faster implementation for String `hasSuffix` & `hasPrefix` & == but for large String arrays.
This is a short yet efficient and to the point implementation. No concurrency.